+++ Localization/TellMeWhen-enUS.lua
@@ -631,6 +631,7 @@ L["AIR"] = "Air"
 L["MUSHROOM"] = "Mushroom %d"
 L["RUNEOFPOWER"] = "Rune %d"
 L["GENERICTOTEM"] = "Totem %d"
+L["GENERICTOTEM_ANY"] = "Any Totem"
 L["RUNES"] = "Rune(s) to check"
 
 
@@ -1830,7 +1831,7 @@ L["GROUPSELECT_TOOLTIP"] = [[|cff7fffffClick|r to edit.
 |cff7fffffClick-and-drag|r to reorder or change domain.]]
 
 L["GROUP_UNAVAILABLE"] = "|TInterface/PaperDollInfoFrame/UI-GearManager-LeaveItem-Transparent:20|t This group cannot be shown due to its overly-restrictive spec/role settings."
-L["GROUP_CANNOT_INTERACTIVELY_POSITION"] = "Could not reposition group because the anchor target is restricted. Use the manual position controls in the group settings."
+L["GROUP_CANNOT_INTERACTIVELY_POSITION"] = "Cannot interactively reposition %s because the anchor target is restricted. Use the manual position controls in the group settings."
 
 --[=[L["CNDT_SLIDER_DESC_BASE"] = [[|cff7fffffMousewheel|r to adjust.
 |cff7fffffShift-Mousewheel|r to adjust x10.