+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v8.7.3
 ### Bug Fixes
 * Fix #1742 - Errors related to improper escaping of user input for the suggestion list.
+* Fixed error `bad argument #1 to 'strlower' (string expected, got boolean)` when using Diminishing Returns icons
 
 ## v8.7.2
 ### Bug Fixes