+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## v8.7.3
+* Added a new "Any Totem" condition that will check all totem slots.
+
 ### Bug Fixes
 * Fix #1742 - Errors related to improper escaping of user input for the suggestion list.
 * Fixed error `bad argument #1 to 'strlower' (string expected, got boolean)` when using Diminishing Returns icons
+* Fix #1755 - Swing Timer conditions with durations other than zero seconds were not triggering updates at the proper moment.
 
 ## v8.7.2
 ### Bug Fixes