+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v8.7.3
 * Added a new "Any Totem" condition that will check all totem slots.
+* Updated totem checking to account for removal of totem APIs in 1.13.3. Totem deaths cannot be accounted for.
 
 ### Bug Fixes
 * Fix #1742 - Errors related to improper escaping of user input for the suggestion list.